TW-1316: Persona external ads. Refactor
